InstaSplit/frontend/requirements.txt
# Streamlit
streamlit==1.41.0

# HTTP requests
requests==2.32.0

# Data manipulation
pandas==2.2.0

# Image handling
Pillow==11.0.0
pillow-heif==0.21.0

# Utilities
python-dotenv==1.0.1
